+++ pkg/frontend/querymiddleware/limits.go
@@ -83,8 +83,11 @@ type Limits interface {
 	// if out of order ingestion is disabled.
 	ResultsCacheTTL(userID string) time.Duration
 
-	// ResultsCacheForOutOfOrderWindowTTL returns TTL for cached results for query that falls into out-of-order ingestion window.
+	// ResultsCacheTTLForOutOfOrderTimeWindow returns TTL for cached results for query that falls into out-of-order ingestion window.
 	ResultsCacheTTLForOutOfOrderTimeWindow(userID string) time.Duration
+
+	// ResultsCacheTTLForCardinalityQuery returns TTL for cached results for cardinality queries.
+	ResultsCacheTTLForCardinalityQuery(userID string) time.Duration
 }
 
 type limitsMiddleware struct {